Implemented custom hook useCopyAttribution to abstract logic from CopyAttributionButton component
