+++ src/components/CopyAttributionButton.jsx
@@ -1,23 +1,9 @@
-import { useEffect, useState } from 'react'
+import { useCopyAttribution } from '@/hooks/useCopyAttribution'
 import { IconCopy, IconCheck } from '@tabler/icons-react'
 import { Button } from '@/components/Button'
 
-const ATTRIBUTION_TEXT = 'Datos extraídos de <a href="https://sueldos.dev/">sueldos.dev</a>'
-
 export function CopyAttributionButton () {
-  const [copied, setCopied] = useState(false)
-
-  useEffect(() => {
-    if (!copied) return
-
-    const timeoutId = setTimeout(() => setCopied(false), 2500)
-
-    return () => clearTimeout(timeoutId)
-  }, [copied])
-
-  const handleCopyAttribution = () => {
-    navigator.clipboard.writeText(ATTRIBUTION_TEXT).then(() => setCopied(true))
-  }
+  const { handleCopyAttribution, copied } = useCopyAttribution()
 
   return (
     <Button onClick={handleCopyAttribution} className='flex gap-2 items-center justify-center my-6 md:mx-auto self-center text-sm'>
